feat: Adds JsonPath interface and event emitter.
Adds JsonPath  interface.
Adds event emitter of JsonPath from child to root parent.

When the user clicks on a segment, an event is emitted.
This event contains information regarding the parent JsonPath and segment data of the segment clicked.

This event is emitted bottom to top, from each recursive child to its parent, finally reaching the root component.

The user can handle this emitted event, and use it according to their needs.